more complete change for both iOS 5 & iOS 6
previous pull request for this actually was relying on whether
respondsToSelector:@selector(shouldAutorotateToInterfaceOrientation:)
would return true … which should have been sufficient to determine
iOS5 vs iOS6 … but the logic wasn't clear that the check wasn't
necessary until after the comparison had been made.

this new logic (and pull request) hopefully make clearer what is
logic to be executed prior to the calls to shouldAutorotate(…) and
what is dependent upon them.